requirements.txt
Django==3.2.19
psycopg2
django-tenant-schemas
